Baby's First Space Race (BFSR) is an instructional Minecraft modpack targeting:
1. Beginner players that don't know anything about Minecraft.
2. Seasons Minecraft players that don't know about mods.
3. Veteran players that want to play a Skinner Box that showers them in prizes.

It is normally hosted at:
http://www.technicpack.net/modpack/babys-first-space-race.616602

The description on the Technic pack site:

"The idea for this game started to support a little team-based Minecraft game
some colleagues and I wanted to play. Most of them had no played Minecraft
before. Having tried technical mods with newcomers like that before, I found
that people needed considerable guidance to know how things even worked in
Minecraft. What to even do when you start a new world is unintuitive. We also
ended up constantly starving to death because everybody would get hurt--which
would deplete the food reserve--while not knowing anything about replenishing
food and farming.

This pack provides some common technical mods, with the Galacticraft moon
mission as something of a final goal. It's unlikely an individual will get to a
rocket in a single session, but it can be possible in a team game over, say, and
afternoon. Its distinguishing features are its quest book and the easing of some
grinding mechanics. The quest book is provided without hardcore mode, goes
through very basic steps in playing with technical mods, and provides rewards.
Grinding is reduced by extremely inflating ore generation, and providing much of
the rote construction as part of quest rewards."