Fix for GitHub issue 5
https://github.com/rockobonaparte/bfsr/issues/5

"The nether chapter has a quest for making obsidian through the Tic smeltery. im assuming its a detection type quest. it requires a smeltery drain and a channel. it detected the channel but not the drain. Just thought id let you know if you don't already"

The drain indeed was not detecting. I could not figure out why. I had to reassert it in HQM. I suspect that this happened when I exported and re-imported the quest book earlier on. Some metadata probably went stupid. Fixed.