Added some more quests on the Progressive Automation branch of the mining tree, and then linked all the mining quests under a common, empty quest that explains the different paths the player can take.
The PA miner now explains how to fully-automate it. I didn't realize that was
possible; I wonder if that showed up since the time I made the quests.

Also, I fixed a goof in the quest book that claimed the chopper worked around
itself, rather than in front of itself. Whoops!