Actually filled in drinking ender quest, and titled flux crystal growing task.
These are some loose ends I never actually finished! Whoops! For the ender stuff,
I had the cup and the straw. It turns out the straw is hard to use, so I just
stuck with the cup as it had the same effect. The straw quest was removed.

The flux crystal growing quest had a generic "Detection Task" title for the last
task. Whoops! Gave it a goofy title ("All Grown Up").