Rigorous edits of the Mekanism quests in the Electrical Power chapter
The smelter quest was ambiguous as to what it actually was. The crusher quest wouldn't detect the crusher (needed NBT-independent detection). I acknowledged redundant Mekanism quests in the post-scarcity chapter. I gated the Mekanism machines on steel and a metallurgic infuser. Finally, I slightly moved around the quests in that chapter to hopefully make the progression look clearer.