Switched tank reward for reactor; workaround for rednet cables; added pun in wither skeleton quest.
The OpenBlocks tanks wouldn't allow one to extract resonant ender with a bucket. It works for other liquids, but not that. It is a mystery. I switched the reward to a hardened tank with 16 buckets of resonant ender.

The rednet cable quest in the reactor section would only detect cables in bulk of 8 at a time, so I just had it detect if it found one cable, and mentioned why I was detecting such a low number.

The Wither Skeleton quest is now known as "Getting Ahead Of The Class" because 'a head' amrite? =D