Got the Philadelphia quest working using an intermediate portal dimension.
I hit a major snag in that I had to build out the ship frame in the right direction, or it doesn't work. That's just too much to ask of players, so I created a 16x16x9 intermediate cell that has an NPC to beam the player into the ship in storage. This also takes care of a customNPC bug where when then ship beams out of space, all the NPCs run around like crazy before disappearing to whatever coordinates they were hard-coded to have. They don't interface well with AE2. The NPCs in the intermediate dimension are just set to stand still.

The quests have been changed, the rewards augmented, the shops updated, and everything. It does, however, need another dry run.