The teleportation code should no longer end up creating Nether portals.
It turned out that the original teleportation code using the base Minecraft teleportation code would create all the portals that would randomly show up. I had to use an alternative method for creating the teleporters. This involves creating a stub teleporter and having the server instance execute the teleportation off of that. I heavily relied on RFTools example code to figure this out!