Ender Ore fixes: nerfed ender ore dimlets; added ender ore veinminer support.
The fixes were made on request of Aciho in issue #7: Ender ore #7

I would not normally care about something being OP in the pack, but if it
reaches a point where people foresake the objectives of it to just do a single
thing, then I will react. Apparently the issue with exploiting ender ore came
up in a server environment, and it was kind of sucking the fun out of some
stuff.